add script for txn upload and group creation
