+++ package.json
@@ -18,7 +18,8 @@
     "db:delete": "node src/scripts/deleteCollections.js",
     "db:reset": "npm run db:delete && npm run db:setup",
     "upload-bundles": "node src/scripts/uploadPremadeBundles.js",
-    "parse-compartment-def": "node src/scripts/parseCompartmentDefinition.js"
+    "parse-compartment-def": "node src/scripts/parseCompartmentDefinition.js",
+    "post-txn-bundles": "node src/scripts/postTransactionBundles.js"
   },
   "repository": {
     "type": "git",
@@ -30,6 +31,7 @@
   },
   "homepage": "https://github.com/projecttacoma/bulk-export-server#readme",
   "dependencies": {
+    "axios": "^1.3.4",
     "bee-queue": "^1.4.0",
     "dotenv": "^10.0.0",
     "fastify": "^3.22.0",
