+++ src/scripts/uploadPremadeBundles.js
@@ -2,6 +2,7 @@ const fs = require('fs');
 const path = require('path');
 const mongoUtil = require('../util/mongo');
 const { createResource } = require('../util/mongo.controller');
+const { createPatientGroupsPerMeasure } = require('../util/groupUtils');
 
 const ecqmContentR4Path = path.resolve(path.join(__dirname, '../../ecqm-content-r4-2021/bundles/measure/'));
 
@@ -52,36 +53,7 @@ const getBundleFiles = (directory, searchPattern) => {
   });
 };
 
-/**
- * Creates a FHIR Group resource that represents the Patients associated with a given Measure
- * and adds it to the database
- * @param {String} measureId An id for the associated FHIR Measure to be used in the Group Id
- * @param {Array} patientIds An array of FHIR Patient ids to be added as Group members
- * @returns {Boolean} True if the Group creation succeeds, false otherwise
- */
-async function createPatientGroupsPerMeasure(measureId, patientIds) {
-  const group = {
-    resourceType: 'Group',
-    id: `${measureId}-patients`,
-    type: 'person',
-    actual: true,
-    member: patientIds.map(pid => ({
-      entity: {
-        reference: `Patient/${pid}`
-      }
-    }))
-  };
 
-  try {
-    await createResource(group, 'Group');
-    return true;
-  } catch (e) {
-    if (e.code !== 11000) {
-      console.log(e.message);
-    }
-    return false;
-  }
-}
 
 /**
  * Uploads all the resources from the specified directory into the
